requirements.txt
asgiref==3.8.1
dependency-injector==4.46.0
Django==5.2
djangorestframework==3.16.0
sqlparse==0.5.3
